pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mpdiff"
version = "0.1.0"
description = "High-dimensional diffusion simulation and Marcenko-Pastur spectral analysis"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "mpdiff contributors" }]
license = { text = "MIT" }
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "matplotlib>=3.8",
  "PyYAML>=6.0",
  "pandas>=2.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "jupyter>=1.0",
  "ruff>=0.5",
  "mypy>=1.8"
]

[project.scripts]
mpdiff = "mpdiff.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
